share HearingTime component with daily docket row and upcoming hearings table row
