+++ client/app/hearings/components/dailyDocket/DailyDocketRowDisplayText.jsx
@@ -1,13 +1,11 @@
-import React from 'react';
-import PropTypes from 'prop-types';
-import { css } from 'glamor';
 import Link from '@department-of-veterans-affairs/caseflow-frontend-toolkit/components/Link';
-import DocketTypeBadge from '../../../components/DocketTypeBadge';
+import PropTypes from 'prop-types';
+import React from 'react';
 
+import { HearingTime } from '../HearingTime';
 import { PreppedCheckbox } from './DailyDocketRowInputs';
 import COPY from '../../../../COPY';
-
-import moment from 'moment-timezone';
+import DocketTypeBadge from '../../../components/DocketTypeBadge';
 
 const hearingPropTypes = PropTypes.shape({
   appealExternalId: PropTypes.string,
@@ -20,28 +18,8 @@ const hearingPropTypes = PropTypes.shape({
   representative: PropTypes.string,
   representativeName: PropTypes.any,
   veteranFileNumber: PropTypes.string,
-  scheduledTimeString: PropTypes.string,
-  regionalOfficeTimezone: PropTypes.string,
-  centralOfficeTimeString: PropTypes.string,
-  readableRequestType: PropTypes.string,
-  regionalOfficeName: PropTypes.string,
-  currentIssueCount: PropTypes.number,
-  paperCase: PropTypes.bool,
-  isVirtual: PropTypes.bool
+  paperCase: PropTypes.bool
 });
-
-export const getDisplayTime = (scheduledTimeString, timezone) => {
-  const val = scheduledTimeString ? moment(scheduledTimeString, 'HH:mm').format('h:mm a') : '';
-
-  if (timezone) {
-    const tz = moment().tz(timezone).
-      format('z');
-
-    return `${val} ${tz}`;
-  }
-
-  return val;
-};
 
 export const getHearingAppellantName = (hearing) => {
   let { appellantFirstName, appellantLastName, veteranFirstName, veteranLastName } = hearing;
@@ -81,48 +59,6 @@ AppellantInformation.propTypes = {
   hearing: hearingPropTypes
 };
 
-const firstParagraphStyle = css({ marginTop: 0 });
-
-const HearingTime = ({ hearing }) => {
-  const localTime = getDisplayTime(
-    hearing.scheduledTimeString,
-    hearing.regionalOfficeTimezone || 'America/New_York'
-  );
-  const coTime = getDisplayTime(hearing.centralOfficeTimeString, 'America/New_York');
-
-  if (hearing.readableRequestType === 'Central') {
-    return (
-      <div>
-        <p {...firstParagraphStyle}>
-          <b>{hearing.readableRequestType}</b>
-        </p>
-        <p>
-          {coTime}<br />
-          {hearing.regionalOfficeName}
-        </p>
-      </div>
-    );
-  }
-
-  return (
-    <div>
-      <p {...firstParagraphStyle}>
-        <b>{hearing.isVirtual ? 'Virtual' : hearing.readableRequestType}</b>
-      </p>
-      <p>
-        {coTime} /<br />
-        {localTime} <br />
-        {hearing.regionalOfficeName}
-      </p>
-      <p>{hearing.currentIssueCount} issues</p>
-    </div>
-  );
-};
-
-HearingTime.propTypes = {
-  hearing: hearingPropTypes
-};
-
 export default class HearingText extends React.Component {
   render () {
     const { hearing, index, user, update, readOnly, initialState } = this.props;
@@ -133,7 +69,7 @@ export default class HearingText extends React.Component {
       </div>
       <div><strong>{index + 1}</strong></div>
       <AppellantInformation hearing={hearing} />
-      <HearingTime hearing={initialState} />
+      <HearingTime hearing={initialState} showIssueCount showRegionalOffice showRequestType />
     </React.Fragment>;
   }
 }
